requirements.txt
anyascii==0.3.2
asgiref==3.8.1
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
defusedxml==0.7.1
Django==5.2.3
django-extensions==4.1
django-filter==25.1
django-modelcluster==6.4
django-permissionedforms==0.1
django-stubs-ext==5.2.1
django-taggit==6.1.0
django-tasks==0.7.0
django-treebeard==4.7.1
djangorestframework==3.16.0
draftjs_exporter==5.1.0
et_xmlfile==2.0.0
filetype==1.2.0
idna==3.10
laces==0.1.2
openpyxl==3.1.5
pillow==11.2.1
pillow_heif==0.21.0
requests==2.32.4
ruff==0.12.0
soupsieve==2.7
sqlparse==0.5.3
telepath==0.3.1
typing_extensions==4.14.0
urllib3==2.5.0
wagtail==7.0.1
Willow==1.10.0
